feat(manage): move to new management system and remove legacy
