--- components/app/account/EmptyAccount.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="flex justify-center p-8">
-    <img src="@/assets/images/account/cuate.svg" alt="" />
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'EmptyAccount',
-}
-</script>
