+++ components/app/header/ManagementOptions.vue
@@ -1,7 +1,7 @@
 <template>
   <!-- logged-in - visualize connected wallet -->
   <div v-if="wallet" class="block xl:flex items-center">
-    <nuxt-link to="/management" class="btn btn-outline bg-base-300">
+    <nuxt-link to="/manage" class="btn btn-outline bg-base-300">
       {{ shrinkAddress(wallet) }}
 
       <jazzicon class="pl-2 w-6 h-6" :address="wallet" :diameter="24" />
