+++ store/account.js
@@ -1,110 +1,96 @@
 import { getAccount, updateProject } from '@/api';
 
+/* Store for managing the associated projects of a connected wallet */
+
 export const state = () => ({
-  projects: null,
+  loading: false,
   wallet: null,
+  projects: [],
 });
 
 export const mutations = {
-  projects(state, payload) {
-    state.projects = payload;
+  reset(state) {
+    state.wallet = null;
+    state.projects = [];
   },
-
-  wallet(state, payload) {
-    state.wallet = payload;
+  loading(state, loading) {
+    state.loading = loading;
+  },
+  wallet(state, wallet) {
+    state.wallet = wallet;
+  },
+  projects(state, projects) {
+    state.projects = projects;
   },
 };
 
 export const actions = {
-  async loadAccount({ commit, dispatch }) {
-    // Reset
+  async softLoad({ state, dispatch }) {
+    if (!state.wallet) {
+      // Only load if account is not loaded
+      await dispatch('load');
+    }
+  },
+
+  async load({ commit, dispatch }) {
     commit('alert/clear', null, { root: true });
+    commit('loading', true);
 
     try {
       const response = await getAccount(this.$axios);
       commit('wallet', response.data.wallet);
       commit('projects', response.data.projects);
     } catch (error) {
-      if (error.response) {
-        const { status } = error.response;
-
-        if (status === 401) {
-          // Authentication failure / timeout
-          await dispatch('auth/timeout', null, { root: true });
-          throw new Error('401 - Unauthorized');
-        }
-
-        if (status === 404) {
-          // No projects available for this account
-          // UI has special component for visualization
-          commit('projects', []);
-          return;
-        }
+      if (error.response?.status === 401) {
+        // Auth timeout
+        await dispatch('auth/timeout', null, { root: true });
+      } else {
+        // Generic error
+        dispatch(
+          'alert/error',
+          { content: this.$i18n.t('general.error.retry'), autoFade: true },
+          {
+            root: true,
+          },
+        );
+        throw new Error(error);
       }
-
-      dispatch('alert/error', 'general.error.retry', { root: true });
-      console.error('Error on backend communication', error);
+    } finally {
+      commit('loading', false);
     }
   },
 
-  async updateProject({ commit, dispatch }, payload) {
-    // Reset
+  async update({ commit, dispatch }, payload) {
     commit('alert/clear', null, { root: true });
-
-    const formData = new FormData();
-
-    Object.entries(payload).forEach((entry) => {
-      const [key, value] = entry;
-
-      if (Array.isArray(value)) {
-        value.forEach((val) =>
-          formData.append(val !== Object(val) ? `${key}[]` : key, val),
-        );
-        if (value.length === 0) {
-          // Workaround for empty arrays
-          formData.append(key, null);
-        }
-      } else if (value instanceof Object) {
-        if (value instanceof File) {
-          formData.append(key, value);
-        } else {
-          Object.keys(value).forEach((objKey) =>
-            formData.append(`${key}[${objKey}]`, value[objKey]),
-          );
-
-          if (Object.keys(value).length === 0) {
-            // Workaround for clearing empty objects. Otherwise it won't be transmitted.
-            formData.append(`${key}[clear]`, '');
-          }
-        }
-      } else {
-        formData.append(key, String(value));
-      }
-    });
+    commit('loading', true);
 
     try {
-      const response = await updateProject(this.$axios, payload.id, formData);
-
+      const response = await updateProject(this.$axios, payload);
       commit('projects', response.data);
       dispatch(
         'alert/success',
-        { content: 'manage.project.changed', autoFade: true },
+        {
+          content: this.$i18n.t('manage.project.changed'),
+          autoFade: true,
+        },
         { root: true },
       );
     } catch (error) {
-      if (error.response) {
-        const { status } = error.response;
-
-        if (status === 401) {
-          // Authentication failure / timeout
-          dispatch('alert/error', 'manage.auth.timeout', { root: true });
-          await this.$router.push('/');
-          return;
-        }
+      if (error.response?.status === 401) {
+        // Auth timeout
+        await dispatch('auth/timeout', null, { root: true });
+      } else {
+        // Generic error
+        dispatch(
+          'alert/error',
+          { content: this.$i18n.t('general.error.retry'), autoFade: true },
+          {
+            root: true,
+          },
+        );
       }
-
-      dispatch('alert/error', 'general.error.retry', { root: true });
-      console.error('Error on project update', error);
+    } finally {
+      commit('loading', false);
     }
   },
 };
